debugging, still need to fix cube camera texture

+++ src/shaders/skybox-fragment.glsl
@@ -1,5 +1,4 @@
-precision mediump float;
-
+precision highp float;
 
 uniform samplerCube u_SamplerCube;
